Move snapshot view icon handling to controller
